Things measured in elements are now in bytes.
This gives a huge speed boost it's not even funny. We avoid the expensive IDIV
in get_elem_size. Turns out it gets called billions of times, while holding a
mutex and otherwise. Also, this lets us remove bufend from the struct. All in
all, we get a 25% improvement. Thank god for gprof.